requirements.txt
fastapi
uvicorn
langchain
openai
twilio
wikipedia
python-decouple
psycopg2-binary
python-multipart
pyngrok
